chore: refactor theming package file structure
