+++ app/client/packages/design-system/theming/src/color/LightModeTheme.ts
@@ -1,4 +1,4 @@
-import { ColorsAccessor } from "../ColorsAccessor";
+import { ColorsAccessor } from "./ColorsAccessor";
 
 import type Color from "colorjs.io";
 import type { ColorTypes } from "colorjs.io/types/src/color";
