+++ app/client/packages/design-system/theming/src/color/index.ts
@@ -1,0 +1,4 @@
+export * from "./ColorsAccessor";
+export * from "./DarkModeTheme";
+export * from "./LightModeTheme";
+export * from "./types";
