--- app/client/packages/design-system/theming/src/components/ThemeContext.tsx
@@ -1,7 +1,0 @@
-import { createContext, useContext } from "react";
-
-export const ThemeContext = createContext<any>(null);
-
-export const useThemeContext = () => {
-  return useContext(ThemeContext);
-};
