+++ app/client/packages/design-system/theming/src/index.ts
@@ -1,6 +1,4 @@
-export { ThemeProvider } from "./components/ThemeProvider";
-export { useThemeContext, ThemeContext } from "./components/ThemeContext";
-export * from "./utils";
-
-export { default as defaultTokens } from "./tokens/defaultTokens.json";
-export { default as themeTokens } from "./tokens/themeTokens.json";
+export * from "./typography";
+export * from "./token";
+export * from "./theme";
+export * from "./color";
