+++ app/client/packages/design-system/theming/src/theme/index.ts
@@ -1,0 +1,3 @@
+export * from "./ThemeContext";
+export * from "./ThemeProvider";
+export * from "./types";
