+++ app/client/packages/design-system/theming/src/token/index.ts
@@ -1,0 +1,4 @@
+export * from "./TokensAccessor";
+
+export { default as defaultTokens } from "./defaultTokens.json";
+export { default as themeTokens } from "./themeTokens.json";
