+++ app/client/packages/design-system/theming/src/typography/index.ts
@@ -1,0 +1,2 @@
+export * from "./typography";
+export * from "./types";
