--- app/client/packages/design-system/theming/src/utils/ColorsAccessor/index.ts
@@ -1,1 +1,0 @@
-export { ColorsAccessor } from "./ColorsAccessor";
