--- app/client/packages/design-system/theming/src/utils/TokensAccessor/index.ts
@@ -1,11 +1,0 @@
-export { DarkModeTheme } from "./DarkModeTheme";
-export { LightModeTheme } from "./LightModeTheme";
-export { TokensAccessor } from "./TokensAccessor";
-export { fontMetricsMap } from "./types";
-
-export type {
-  ThemeTokens,
-  TokenSource,
-  TypographyVariants,
-  Typography,
-} from "./types";
