--- app/client/packages/design-system/theming/src/utils/index.ts
@@ -1,8 +1,0 @@
-export * from "./TokensAccessor";
-export * from "./ColorsAccessor";
-
-export type {
-  ThemeTokens,
-  TokenSource,
-  TypographyVariants,
-} from "./TokensAccessor";
